moved compare bar into home components
