+++ src/components/homeComponents/CompareBar/CoinSelector.tsx
