+++ src/components/homeComponents/CompareBar/index.tsx
